arm: Fix switch tables for thumb-1 with -mpure-code [PR96768]
In comment 14 from PR94538, it was suggested to switch off jump tables
on thumb-1 cores when using -mpure-code, like we already do for thumb-2.

This is what this patch does, and also restores the previous value of
CASE_VECTOR_PC_RELATIVE since it was not the right way of handling
this.

It also adds a new test, since the existing no-casesi.c did not catch
this problem.

Tested by running the whole testsuite with -mpure-code -mcpu=cortex-m0
-mfloat-abi=soft, no regression and the new test passes (and fails
without the fix).

2020-08-27  Christophe Lyon  <christophe.lyon@linaro.org>

	gcc/
	* config/arm/arm.h (CASE_VECTOR_PC_RELATIVE): Remove condition on
	-mpure-code.
	* config/arm/thumb1.md (tablejump): Disable when -mpure-code.

	gcc/testsuite/
	* gcc.target/arm/pure-code/pr96768.c: New test.